Changed Message server_time to a Double.

+++ TinyChat/Message.swift
@@ -11,7 +11,7 @@ import Foundation
 struct Message: Codable {
     var msg: String
     var client_time:Double
-    var server_time:Int?
+    var server_time:Double?
     
     init(msg: String, client_time: Double) {
         self.msg = msg
@@ -19,7 +19,7 @@ struct Message: Codable {
         self.server_time = nil
     }
     
-    init(msg: String, client_time: Double, server_time: Int) {
+    init(msg: String, client_time: Double, server_time: Double) {
         self.msg = msg
         self.client_time = client_time
         self.server_time = server_time
